+++ consulacl/common.go
@@ -1,0 +1,17 @@
+package consulacl
+
+const FieldName = "name"
+const FieldToken = "token"
+const FieldAccessor = "accessor"
+const FieldSecret = "secret"
+const FieldType = "type"
+
+const FieldRule = "rule"
+
+const FieldScope = "scope"
+const FieldPrefix = "prefix"
+const FieldPolicy = "policy"
+
+const FieldDescription = "description"
+const FieldPolicies = "policies"
+const FieldLocal = "local"